requirements.txt
numpy==1.15.0
torchvision==0.2.2.post3
torch==1.0.1.post2
Pillow==6.1.0
